feat: Refactor ERP macros to versioned structure
- Updated import paths for ERP macros to a new versioned directory structure (v1).
- Added for ERPAliMacro, ERPBrandiMacro, ERPEtcSiteMacro, ERPZigzagMacro, and ERPGmaAucMacro in the v1 directory.
- Ensured compatibility with existing functionality while organizing code for better maintainability.
